Section 12 completed

+++ eShopSource/Services/Basket/Basket.API/Models/ShoppingCart.cs
@@ -4,7 +4,7 @@ public class ShoppingCart
 {
     public string UserName { get; set; } = default!;
     public List<ShoppingCartItem> Items { get; set; } = new();
-    public decimal TotalPrice => Items.Sum(i => i.Price * i.Quantity);
+    public double TotalPrice => Items.Sum(i => i.Price * i.Quantity);
 
     public ShoppingCart(string userName)
     {
